(Should) fix amenities not showing up in onboarding
